requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-dotenv==1.0.0
requests==2.28.2
google-generativeai==0.3.1
gunicorn==20.1.0